# internal/module/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Core Ledger API",
    "description": "API documentation for Core Ledger system - COA Account Management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@core-ledger.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.core-ledger.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/rule-category/coa-rules": {
      "get": {
        "tags": ["rule-category"],
        "summary": "Get COA account rules",
        "description": "Get all COA account rules structure with types, groups, and steps for generating account codes",
        "operationId": "getCoaAccountRules",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoaAccountRulesResponse"
                },
                "example": {
                  "data": [
                    {
                      "id": 1,
                      "code": "ASSET",
                      "name": "Asset",
                      "separator": ":",
                      "groups": [
                        {
                          "id": 1,
                          "code": "FLOAT",
                          "name": "Float",
                          "input_type": "SELECT",
                          "separator": ":",
                          "steps": [
                            {
                              "step_id": 1,
                              "step_order": 1,
                              "type": "SELECT",
                              "label": "",
                              "category_code": "CURRENCY",
                              "values": [
                                {
                                  "id": 22,
                                  "name": "VND",
                                  "value": "VND"
                                },
                                {
                                  "id": 23,
                                  "name": "USD",
                                  "value": "USD"
                                }
                              ],
                              "input_type": "SELECT",
                              "separator": "."
                            }
                          ]
                        }
                      ]
                    }
                  ],
                  "error": null
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CoaAccountRulesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoaAccountRuleTypeResp"
            }
          },
          "error": {
            "nullable": true,
            "$ref": "#/components/schemas/ResponseError"
          }
        }
      },
      "CoaAccountRuleTypeResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string",
            "example": "ASSET"
          },
          "name": {
            "type": "string",
            "example": "Asset"
          },
          "separator": {
            "type": "string",
            "example": ":"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoaAccountRuleGroupResp"
            }
          }
        }
      },
      "CoaAccountRuleGroupResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string",
            "example": "FLOAT"
          },
          "name": {
            "type": "string",
            "example": "Float"
          },
          "input_type": {
            "type": "string",
            "enum": ["SELECT", "TEXT"],
            "example": "SELECT"
          },
          "separator": {
            "type": "string",
            "example": ":"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoaAccountRuleStepResp"
            }
          }
        }
      },
      "CoaAccountRuleStepResp": {
        "type": "object",
        "properties": {
          "step_id": {
            "type": "integer",
            "format": "int64"
          },
          "step_order": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": ["SELECT", "TEXT"],
            "example": "SELECT"
          },
          "label": {
            "type": "string",
            "example": "Details"
          },
          "category_code": {
            "type": "string",
            "nullable": true,
            "example": "CURRENCY"
          },
          "values": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/RuleValueResp"
            }
          },
          "input_code": {
            "type": "string",
            "nullable": true,
            "example": "DETAILS"
          },
          "input_type": {
            "type": "string",
            "nullable": true,
            "example": "TEXT"
          },
          "separator": {
            "type": "string",
            "example": "."
          }
        }
      },
      "RuleValueResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "VND"
          },
          "value": {
            "type": "string",
            "example": "VND"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ResponseError"
          }
        }
      },
      "ResponseError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "property": {
            "type": "string",
            "nullable": true,
            "example": "account_no"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}